scrapers/loblaws/requirements.txt
requests_html==0.10.0
lxml==4.9.2
pytest==7.2.1
aiohttp==3.8.3
aiolimiter==1.0.0
firebase-admin==6.1.0
python-dotenv==0.21.1